rpma: common: separate client_init
