+++ engines/librpma.c
@@ -56,26 +56,14 @@ struct client_data {
 
 static int client_init(struct thread_data *td)
 {
-	struct librpma_common_client_options *o = td->eo;
 	struct librpma_common_client_data *ccd;
 	struct client_data *cd;
-	struct ibv_context *dev = NULL;
-	struct rpma_conn_cfg *cfg = NULL;
-	char port_td[LIBRPMA_COMMON_PORT_STR_LEN_MAX];
-	struct rpma_conn_req *req = NULL;
-	enum rpma_conn_event event;
+	unsigned int sq_size;
 	uint32_t cq_size;
-	struct rpma_conn_private_data pdata;
-	struct librpma_common_workspace *ws;
-	size_t server_mr_size;
+	struct rpma_conn_cfg *cfg = NULL;
 	int remote_flush_type;
 	struct rpma_peer_cfg *pcfg = NULL;
-	unsigned int sq_size;
-	int ret = 1;
-
-	/* configure logging thresholds to see more details */
-	rpma_log_set_threshold(RPMA_LOG_THRESHOLD, RPMA_LOG_LEVEL_INFO);
-	rpma_log_set_threshold(RPMA_LOG_THRESHOLD_AUX, RPMA_LOG_LEVEL_INFO);
+	int ret;
 
 	/* not supported readwrite = trim / randtrim / trimwrite */
 	if (td_trim(td)) {
@@ -94,41 +82,6 @@ static int client_init(struct thread_data *td)
 		td_verror(td, errno, "calloc");
 		free(ccd);
 		return 1;
-	}
-
-	/* allocate all in-memory queues */
-	ccd->io_us_queued = calloc(td->o.iodepth, sizeof(struct io_u *));
-	if (ccd->io_us_queued == NULL) {
-		td_verror(td, errno, "calloc");
-		goto err_free_cd;
-	}
-
-	ccd->io_us_flight = calloc(td->o.iodepth, sizeof(struct io_u *));
-	if (ccd->io_us_flight == NULL) {
-		td_verror(td, errno, "calloc");
-		goto err_free_io_us_queued;
-	}
-
-	ccd->io_us_completed = calloc(td->o.iodepth, sizeof(struct io_u *));
-	if (ccd->io_us_completed == NULL) {
-		td_verror(td, errno, "calloc");
-		goto err_free_io_us_flight;
-	}
-
-	/* obtain an IBV context for a remote IP address */
-	ret = rpma_utils_get_ibv_context(o->hostname,
-				RPMA_UTIL_IBV_CONTEXT_REMOTE,
-				&dev);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_utils_get_ibv_context");
-		goto err_free_io_us_completed;
-	}
-
-	/* create a connection configuration object */
-	ret = rpma_conn_cfg_new(&cfg);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_conn_cfg_new");
-		goto err_free_io_us_completed;
 	}
 
 	/*
@@ -173,6 +126,13 @@ static int client_init(struct thread_data *td)
 	}
 	cq_size = sq_size;
 
+	/* create a connection configuration object */
+	ret = rpma_conn_cfg_new(&cfg);
+	if (ret) {
+		librpma_td_verror(td, ret, "rpma_conn_cfg_new");
+		goto err_free_cd;
+	}
+
 	/* apply queue sizes */
 	ret = rpma_conn_cfg_set_sq_size(cfg, sq_size);
 	if (ret) {
@@ -185,67 +145,13 @@ static int client_init(struct thread_data *td)
 		goto err_cfg_delete;
 	}
 
-	/* create a new peer object */
-	ret = rpma_peer_new(dev, &ccd->peer);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_peer_new");
+	if ((ret = librpma_common_client_init(td, ccd, cfg)))
 		goto err_cfg_delete;
-	}
-
-	/* create a connection request */
-	if ((ret = librpma_common_td_port(o->port, td, port_td)))
-		goto err_peer_delete;
-	ret = rpma_conn_req_new(ccd->peer, o->hostname, port_td, cfg, &req);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_conn_req_new");
-		goto err_peer_delete;
-	}
-
-	ret = rpma_conn_cfg_delete(&cfg);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_conn_cfg_delete");
-		goto err_peer_delete;
-	}
-
-	/* connect the connection request and obtain the connection object */
-	ret = rpma_conn_req_connect(&req, NULL, &ccd->conn);
-	if (ret) {
-		librpma_td_verror(td, ret, "rpma_conn_req_connect");
-		goto err_req_delete;
-	}
-
-	/* wait for the connection to establish */
-	ret = rpma_conn_next_event(ccd->conn, &event);
-	if (ret) {
-		goto err_conn_delete;
-	} else if (event != RPMA_CONN_ESTABLISHED) {
-		log_err(
-			"rpma_conn_next_event returned an unexptected event: (%s != RPMA_CONN_ESTABLISHED)\n",
-			rpma_utils_conn_event_2str(event));
-		goto err_conn_delete;
-	}
-
-	/* get the connection's private data sent from the server */
-	if ((ret = rpma_conn_get_private_data(ccd->conn, &pdata)))
-		goto err_conn_delete;
-
-	ws = pdata.ptr;
-
-	/* create the server's memory representation */
-	if ((ret = rpma_mr_remote_from_descriptor(&ws->descriptors[0],
-			ws->mr_desc_size, &ccd->server_mr)))
-		goto err_conn_delete;
-
-	/* get the total size of the shared server memory */
-	if ((ret = rpma_mr_remote_get_size(ccd->server_mr, &server_mr_size))) {
-		librpma_td_verror(td, ret, "rpma_mr_remote_get_size");
-		goto err_conn_delete;
-	}
 
 	/* get flush type of the remote node */
 	if ((ret = rpma_mr_remote_get_flush_type(ccd->server_mr, &remote_flush_type))) {
 		librpma_td_verror(td, ret, "rpma_mr_remote_get_flush_type");
-		goto err_conn_delete;
+		goto err_cleanup_common;
 	}
 
 	cd->flush_type = (remote_flush_type & RPMA_MR_USAGE_FLUSH_TYPE_PERSISTENT) ?
@@ -256,54 +162,40 @@ static int client_init(struct thread_data *td)
 		ret = rpma_peer_cfg_new(&pcfg);
 		if (ret) {
 			librpma_td_verror(td, ret, "rpma_peer_cfg_new");
-			goto err_conn_delete;
+			goto err_cleanup_common;
 		}
 
 		ret = rpma_peer_cfg_set_direct_write_to_pmem(pcfg, true);
 		if (ret) {
 			librpma_td_verror(td, ret, "rpma_peer_cfg_set_direct_write_to_pmem");
-			goto peer_cfg_delete;
+			goto err_cleanup_common;
 		}
 
 		ret = rpma_conn_apply_remote_peer_cfg(ccd->conn, pcfg);
 		if (ret) {
 			librpma_td_verror(td, ret, "rpma_conn_apply_remote_peer_cfg");
-			goto peer_cfg_delete;
+			goto err_cleanup_common;
 		}
 
 		(void) rpma_peer_cfg_delete(&pcfg);
 	}
 
-	ccd->ws_size = server_mr_size;
 	ccd->client_data = cd;
 	td->io_ops_data = ccd;
 
 	return 0;
 
-peer_cfg_delete:
-	(void) rpma_peer_cfg_delete(&pcfg);
-
-err_conn_delete:
+err_cleanup_common:
+	/* XXX to be replaced with librpma_common_client_cleanup */
 	(void) rpma_conn_disconnect(ccd->conn);
 	(void) rpma_conn_delete(&ccd->conn);
-
-err_req_delete:
-	if (req)
-		(void) rpma_conn_req_delete(&req);
-err_peer_delete:
 	(void) rpma_peer_delete(&ccd->peer);
+	free(ccd->io_us_queued);
+	free(ccd->io_us_flight);
+	free(ccd->io_us_completed);
 
 err_cfg_delete:
 	(void) rpma_conn_cfg_delete(&cfg);
-
-err_free_io_us_completed:
-	free(ccd->io_us_completed);
-
-err_free_io_us_flight:
-	free(ccd->io_us_flight);
-
-err_free_io_us_queued:
-	free(ccd->io_us_queued);
 
 err_free_cd:
 	free(cd);
